+++ pyogrio/core.py
@@ -180,6 +180,8 @@ def read_info(path_or_buffer, /, layer=None, encoding=None, **kwargs):
                 "geometry": "<geometry type>",
                 "features": <feature count>,
                 "driver": "<driver>",
+                "dataset_metadata" "<dict of dataset metadata or None>"
+                "layer_metadata" "<dict of layer metadata or None>"
             }
     """
     path, buffer = get_vsi_path(path_or_buffer)